requirements.txt
py-cord[voice]
PyNaCl
googlesearch-python==1.2.3
aiomysql
chat-exporter
art
cryptography
bs4
validators
captcha
spotipy
lavalink
forex_python
googletrans==3.1.0a0
akinator.py
openai
discord_webhook
Pillow==9.5.0
qrcode